Fixed a few compiler warnings
Hey, so first off, thanks for the great library!  It's super-cool and super-simple, exactly what I was looking for.  I'm using it in a project and compiling with -Wextra.  This led to 3 very minor fixes that I've added in this commit:

* There were 2 places where you had the second and third parameters of memset() backwards (the zeroes are the second parameter, not the third).
* And the other one was a default value for a switch-statement.

Anyway, thanks again for the awesome tool and hope this helps. :)